Fixes #31, fully rewrote permissions and checks.
Moved to a better permission check system that is part of the framework.
Added Policy and Request for Activity model. Policy should deal with
permission checks, keeping those out of the model itself. Request should
deal with validation and permission check for cases of storage, both
create and update are affected.