Closes #8 and #9, solutions will be applied as needed.
These are ust tryouts for determining the closest marker and opening
InfoWindows. Current solution would let user DoubleClick on a location,
that places the marker and determines if that is the colsest one,
closest markers get automatic enabling of InfoWindow. This code does
solve both issues, although the concrete implementation will depend on
the game (this will only happen later in development).