Changes for the question inputs of Activity Item.
Added visuals for question types data input (the data is not stored
yet). Chnaged the JS a bit and am now using the Vue.js and Components
for this case. Managing multiple input elements acting as one and having
certain visuals with plain JS turned out to be a bit too dirty. Currrent
solution is a bit more complex (using Vue.js), but has a more logical
code and structure.