References #18, implemented ActivityItem update.
ActivityItem can now be updated. The the backend is fully implemented to
handle all the cases as needed. The code migth need a serious
refactoring as it has quite a few repetitive code (it might be
sufficient to move the repetitive fragments into standalone methods; it
mihgt also be good to use the same code for both create/and update
cases, but that would raise the complexity). Also fixed a few small bugs
for the front-end.