References #18, implemented deletion.
Deletion is possible for both Activity and ActivityItems. In both cases
the system also takes care of removing the uploaded files. The removal
is not using forms, but allows JS code to deal with that (creates form
and ausbmits it). The solution should be applicable everywhere and only
required a few "data" attributes to be defined.