Refernces #36, #26, #25, implemented new gameplay.
This time it is based on Vue.js and is able to display all the
information available. Some of the question types are fully ready to be
answered, some are sill just showing data (the only one really static
question type is MatchPairs, it needs to be decided how that one should
be). Game currently does not store anything and it also allows to access
all the questions, without distance-based limitation (this is needed for
ease of testing and will change soon).