References #36 and #42, color changes, also fixed a bug.
Made sure that answered markers are displayed with green icon and do not
open a question dialog. Also added a functionality to connect markers
with arrowed lines (this one is not being used as it requires a
configuration option to activate). Fixed a bug with MatchPairs, moved
shuffling code to dialog being opened as it is now mounted just once.