Referencres #36, visualtunes and improvements for the gameplay.
Moved textarea after embedded content. Made sure that OneCorrentAnswer
and MultipleCorrectAnswers questions allow marking an answer by clicking
on the whole row. Tuned MatchPairs to display visually distinguishable
boxes, added minimum height (a solutio to set the height to the same
value is required). Also moved some of the inline styles to stylesheets
file (this has to be moved in full).