References #36 and #41 added results dialog.
Added a dialog that can not be closed and is displayed after the game
has finished. This is currently just listing all the questions, but it
should be displaying the stats for answers given. Added a logic to
determine if the game has finished (last question has been answered).
Also changed the way matched pairs are displayed, the elements have
their opacity changed so that they would visually differ.