References #36, refactored and chaned logic a bit.
Implemented a way better logic for game ending detection and passing of
data. Removed the setTimeout code and set the value once the dialog is
closed (this would also allow to show results to the player and only
notify the parent of results when needed; thus only ending the game once
it makes sense to). Also added a logic to check if al the questions have
been answered, new logic allows it to have answers to questions that no
longer exist, these would now be ignored as comparison is not based on
just length. Also added a baseUrl parameter logic for the ajax calls.