Resolves #50, added reCaptcha support to registration.
Also added a small CSS fix for width on devices with small screens
(smaller phones). Please note the additions made to .env.example file
and move those into an existing one and fill with data. Make sure to run
the "composer install" or update command to install the required
dependencies.