Changed logic of attaching Items.
Updated Sortable library a bit (this solved a few issues on it choking
all of a sudden). Also makde sure the whole items are draggable, not
just a certain handle element. Had to disable the tooltip as it was
misbehaving rather bad (it seems to be moved to the new item with
cloning and then has other issues).