References #47, most of the logic implemented.
Most of the searhc logic is in place. Same goes for visuals. Some
details will still need to be handled and solution should be made mobile
friendly (some parts of it at least). Search is fully functional, any
unused data options are disabled to signal the fact that those will be
implemented later on. NB! An additional JS module was added, requires
running "npm install".