References #47, should be close to fully functional.
A few condition types are still disabled, but that is due to fields
still not being present at the certain content type. Should be rather
mobile friendly, with hiding some of the information on the extra small
screens. Needs some testing, especiallt with regard to text search
(current approach is using LIKE with the whole text value provided,
without trying to split it into keywords).