References #64, added initial implementation.
The results pages should be fully functional and working well enough
with mobile devices. The issue is with the need of optimization
(especially limiting the number of queries and forcing preload of some
related data to keep it fast enough een if the volumes grow). Another
change would be to display limited number of results, according to some
logic. Displaying all might raise issues with page loads becoming too
slow or allowed memory size exhausted.