References #71, moved options into services.
Also rewrote the code. Templates are now almost free from additional
code clutter and controllers are mostly responsible for passing the data
in. Also did some refactoring. This one still needs a bit of testing and
refactoring to the options classes (most of them have one method in
common; good solution might be to create a base class that is
inherited).