Refreences #53, dealt with one correct answer.
Only dealt with One Correct Answer for now. Multiple Correct Answers should be dealt with soon enough. Also changed the logic of minimum required options, the number has been set to 2 (this logically leaves one of those to be corrent and one incorrect; this works best with single choice one, but still should also be applicable to the other one).