Major rewrite to image processing.
Moved the processing code to a standalone service and did a lot of
recatroing. Made sure that service is registered as a singleton and
included it as needed. It could need a bit more refactoring in the
future. Made sure that Image object of Intervention is destroyed once
the processing is done. This should keep the memory footprint at bay, as
much as it can be.