Resolves #33, added file type check.
System will check the image type and would only consider the acceptable
ones. In case an unacceptable file is chosen, the input is reset. There
is one gotcha: if previous image existed and newly selected file was not
suitable, the form will not show the data for old image, yet the image
for the option will be kept. That is not a major issue and could be delt
with later.