Resolves #74, added soft deletion.
Added SoftDeletes trait to Activity model. Changed the code to account
for that (lsting of current games for the user and single game view).
Also changed the deletion code within the controller. It uses a hybrid
approach by only applying the soft deletion to activities with games. If
no games are present, activity really gets deleted. This change requires
running a database migrations.