References #139, added new images.
Added the new images and made sure the system would use the SVG files
by default. Added the PNG counterparts for the new files so that Mozilla
Backpack would work. Alos added a versioning buster component to image
URL, this should help with getting new images and not the cached
versions (could also prove to be useful inthe future).