References #137, added guidelines modal.
Added the modal itself with all the looks and feel. This changes the
current logic and would need additional modifications to fit in
organically. It might be wise to either show this one or the game
information modal. This should mostly make it for the time requiredby
the positioning. The case of already completed game should change. It
does not need positioning or anything else. It should just show the
results modal based on the current game data. It will take some trial
and error to get it done right.