Changed game dialogs.
Moved most of the dialog components outside of the GameMap. This allows
showing results without the need to load maps or ask for geolocation.
This shortens the load times significantly. Most of the other dialogs
are now needed before the game loads, this makes the main root element a
perfect place for those. There is a need to sort out which dialog is
being shown, the general one with tips or one with game information. It
might make sense to show the one with tips just once per device.
Currently the one with tips is always shown.