Resolves #137, fixed animation glitch.
Had to resort to setting the minimum height on the tips container
element so that it would not be suddenly resized for a split second and
produce a jumpy glitch effect. The solution coul have used the style
binding instead, but that seems to have issues in the current version of
Vue.