Resolves #153, implemented qr code for Activities.
NB! A new library is used for doing the heavy lifting of the QR Code
related functionality. Please make sure to run the 'composer update'
command. The functionality is generan enough (on the front-end side) and
could be used to cover any other use cases. The only thing that is
needed it to pass the API endpoint to call.