Resolves #161, implemented date filtering possibility.
Added inputs for date filtering possibility. Used the datetime-local
input types (might now be suppored by all the browsers), but it should
be good enough for now. In case it is not available, user will be able
to enter a textual representation of a date and time string, this will
be parsed on the backend. If parsing in not successful, then the value
will be assigned to NULL and parsing will silently fail. Also fixed an
issue of a few texts not being made translatable. In case better input
that works in all browsers is required, one could be added later on.