Changed Activity validator, made playing_time required.
If it is not required and is left empty, the value sent into the
database is incompatible with the column definition (a positive integer
value is required). This is not a big deal, but could still happen.
Validator notice (error) is way better than a meaningless blank screen
of with code 500.