Resolves #188, added modal dialog.
A modal dialog will be shown to the unanthenticated users when pressing
the play button. The dialog shows a warning text and offers to eiter
proceed and start the game or close the dialog and create an account.
The latter one is not part of the flow and should be done by the user.