Resolves #194, implemented exiting to url.
Added logic to allow passing in the URL that would be used when game is
left. It defaults to listing of all activities. It also has an
additional protection of making sure that only URLs local to the
platform are acceptable. All the query filters are retained for the
listings.