Resolves #203, added icon and logic.
The icons will indicate if current user is authenticated or not. The
icon for not-authenticated user will be shown as grey, instead of
standard black. Clicking on an icon will show a tooltip containing
either currently logged in user name or a text that the user has not
logged in.