Should fix an edge case error.
It seems that in rare cases the initialization could fail as marker icon
state detection code is unable to fetch the question object, although
the data should be present. Added a nextTick to triggering of that
function and moved the position listener code to be initialized after
all the markers have been dealt with. This should make sure that data is
present and the Vue had the time to deal with that.